refactor: club api 함수 파일별로 분리
