+++ src/entities/club/api/deleteClub.ts
@@ -1,0 +1,6 @@
+import { instance } from "@/shared/api/instance";
+
+export const deleteClub = async (clubId: number) => {
+  const { data } = await instance.delete(`/clubs/${clubId}`);
+  return data;
+};
