+++ src/entities/club/api/patchClubApproval.ts
@@ -1,0 +1,9 @@
+import { instance } from "@/shared/api/instance";
+
+export const patchClubApproval = async (
+  clubId: number,
+  body: { approved: boolean },
+) => {
+  const { data } = await instance.patch(`/clubs/${clubId}/approval`, body);
+  return data;
+};
